annotation type for a function that takes a float and return the floor of it

+++ 0x00-python_variable_annotations/2-floor.py
@@ -5,8 +5,8 @@ Module Docs
 import math
 
 
-def floor(n: float) -> float:
+def floor(n: float) -> int:
     """
     Function Docs
     """
-    return math.floor(n)
+    return math.floor(int(n))
